refactor(ui): extract shared RouteTabStrip from machine/collection tab strips (PP-slrd.1)
MachineTabStrip and CollectionTabStrip were ~95% identical — the overflow
detection, ResizeObserver, active-tab scrollIntoView centering, edge-fade
gradient, and nav markup were verbatim copies. Extract a RouteTabStrip that
owns all of that and takes a tab list + badge config; both strips become thin
configs. testids are preserved (index tab keys off its label so
machine-tab-info / collection-tab-overview are unchanged). A future
/c/tag/[slug] source is the third consumer.
